Move assets to different folders according to their route
