+++ src/components/home/HowToUse.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from 'next/image';
-import logo from '../../../public/assets/logo1.png';
+import logo from '../../../public/assets/images/logo1.png';
 
 const steps = [
   {
